refactor: all task api handlers moved to controllers
all task related controllers moved to controllers package
