+++ backend/controllers/complete_task.go
@@ -1,0 +1,50 @@
+package controllers
+
+import (
+	"ccsync_backend/models"
+	"ccsync_backend/utils/tw"
+	"encoding/json"
+	"fmt"
+	"io"
+	"net/http"
+)
+
+func CompleteTaskHandler(w http.ResponseWriter, r *http.Request) {
+	if r.Method == http.MethodPost {
+		body, err := io.ReadAll(r.Body)
+		if err != nil {
+			http.Error(w, fmt.Sprintf("error reading request body: %v", err), http.StatusBadRequest)
+			return
+		}
+		defer r.Body.Close()
+
+		fmt.Printf("Raw request body: %s\n", string(body))
+
+		var requestBody models.CompleteTaskRequestBody
+
+		err = json.Unmarshal(body, &requestBody)
+		if err != nil {
+			http.Error(w, fmt.Sprintf("error decoding request body: %v", err), http.StatusBadRequest)
+			return
+		}
+
+		email := requestBody.Email
+		encryptionSecret := requestBody.EncryptionSecret
+		uuid := requestBody.UUID
+		taskuuid := requestBody.TaskUUID
+
+		if taskuuid == "" {
+			http.Error(w, "taskuuid is required", http.StatusBadRequest)
+			return
+		}
+
+		if err := tw.CompleteTaskInTaskwarrior(email, encryptionSecret, uuid, taskuuid); err != nil {
+			http.Error(w, err.Error(), http.StatusInternalServerError)
+			return
+		}
+
+		http.Redirect(w, r, "/tasks", http.StatusSeeOther)
+		return
+	}
+	http.Error(w, "Invalid request method", http.StatusMethodNotAllowed)
+}
