+++ backend/main.go
@@ -2,12 +2,7 @@ package main
 
 import (
 	"ccsync_backend/controllers"
-	"ccsync_backend/models"
-	"ccsync_backend/utils/tw"
 	"encoding/gob"
-	"encoding/json"
-	"fmt"
-	"io"
 	"log"
 	"net/http"
 	"os"
@@ -54,11 +49,11 @@ func main() {
 	mux.HandleFunc("/api/user", app.UserInfoHandler)
 	mux.HandleFunc("/auth/logout", app.LogoutHandler)
 	mux.HandleFunc("/tasks", app.TasksHandler)
-	mux.HandleFunc("/add-task", AddTaskHandler)
-	mux.HandleFunc("/edit-task", EditTaskHandler)
-	mux.HandleFunc("/modify-task", ModifyTaskHandler)
-	mux.HandleFunc("/complete-task", CompleteTaskHandler)
-	mux.HandleFunc("/delete-task", DeleteTaskHandler)
+	mux.HandleFunc("/add-task", controllers.AddTaskHandler)
+	mux.HandleFunc("/edit-task", controllers.EditTaskHandler)
+	mux.HandleFunc("/modify-task", controllers.ModifyTaskHandler)
+	mux.HandleFunc("/complete-task", controllers.CompleteTaskHandler)
+	mux.HandleFunc("/delete-task", controllers.DeleteTaskHandler)
 
 	log.Println("Server started at :8000")
 	if err := http.ListenAndServe(":8000", app.EnableCORS(mux)); err != nil {
@@ -66,205 +61,3 @@ func main() {
 	}
 }
 
-func DeleteTaskHandler(w http.ResponseWriter, r *http.Request) {
-	if r.Method == http.MethodPost {
-		body, err := io.ReadAll(r.Body)
-		if err != nil {
-			http.Error(w, fmt.Sprintf("error reading request body: %v", err), http.StatusBadRequest)
-			return
-		}
-		defer r.Body.Close()
-
-		var requestBody models.DeleteTaskRequestBody
-
-		err = json.Unmarshal(body, &requestBody)
-		if err != nil {
-			http.Error(w, fmt.Sprintf("error decoding request body: %v", err), http.StatusBadRequest)
-			return
-		}
-
-		email := requestBody.Email
-		encryptionSecret := requestBody.EncryptionSecret
-		uuid := requestBody.UUID
-		taskuuid := requestBody.TaskUUID
-
-		if taskuuid == "" {
-			http.Error(w, "taskuuid is required", http.StatusBadRequest)
-			return
-		}
-
-		if err := tw.DeleteTaskInTaskwarrior(email, encryptionSecret, uuid, taskuuid); err != nil {
-			http.Error(w, err.Error(), http.StatusInternalServerError)
-			return
-		}
-		http.Redirect(w, r, "/tasks", http.StatusSeeOther)
-		return
-	}
-	http.Error(w, "Invalid request method", http.StatusMethodNotAllowed)
-}
-
-func CompleteTaskHandler(w http.ResponseWriter, r *http.Request) {
-	if r.Method == http.MethodPost {
-		body, err := io.ReadAll(r.Body)
-		if err != nil {
-			http.Error(w, fmt.Sprintf("error reading request body: %v", err), http.StatusBadRequest)
-			return
-		}
-		defer r.Body.Close()
-
-		fmt.Printf("Raw request body: %s\n", string(body))
-
-		var requestBody models.CompleteTaskRequestBody
-
-		err = json.Unmarshal(body, &requestBody)
-		if err != nil {
-			http.Error(w, fmt.Sprintf("error decoding request body: %v", err), http.StatusBadRequest)
-			return
-		}
-
-		email := requestBody.Email
-		encryptionSecret := requestBody.EncryptionSecret
-		uuid := requestBody.UUID
-		taskuuid := requestBody.TaskUUID
-
-		if taskuuid == "" {
-			http.Error(w, "taskuuid is required", http.StatusBadRequest)
-			return
-		}
-
-		if err := tw.CompleteTaskInTaskwarrior(email, encryptionSecret, uuid, taskuuid); err != nil {
-			http.Error(w, err.Error(), http.StatusInternalServerError)
-			return
-		}
-
-		http.Redirect(w, r, "/tasks", http.StatusSeeOther)
-		return
-	}
-	http.Error(w, "Invalid request method", http.StatusMethodNotAllowed)
-}
-
-func EditTaskHandler(w http.ResponseWriter, r *http.Request) {
-	if r.Method == http.MethodPost {
-		body, err := io.ReadAll(r.Body)
-		if err != nil {
-			http.Error(w, fmt.Sprintf("error reading request body: %v", err), http.StatusBadRequest)
-			return
-		}
-		defer r.Body.Close()
-
-		fmt.Printf("Raw request body: %s\n", string(body))
-
-		var requestBody models.EditTaskRequestBody
-
-		err = json.Unmarshal(body, &requestBody)
-		if err != nil {
-			http.Error(w, fmt.Sprintf("error decoding request body: %v", err), http.StatusBadRequest)
-			return
-		}
-
-		email := requestBody.Email
-		encryptionSecret := requestBody.EncryptionSecret
-		uuid := requestBody.UUID
-		taskuuid := requestBody.TaskUUID
-		description := requestBody.Description
-
-		if taskuuid == "" {
-			http.Error(w, "taskuuid is required", http.StatusBadRequest)
-			return
-		}
-
-		if err := tw.EditTaskInTaskwarrior(uuid, description, email, encryptionSecret, taskuuid); err != nil {
-			http.Error(w, err.Error(), http.StatusInternalServerError)
-			return
-		}
-
-		http.Redirect(w, r, "/tasks", http.StatusSeeOther)
-		return
-	}
-	http.Error(w, "Invalid request method", http.StatusMethodNotAllowed)
-}
-
-func ModifyTaskHandler(w http.ResponseWriter, r *http.Request) {
-	if r.Method == http.MethodPost {
-		body, err := io.ReadAll(r.Body)
-		if err != nil {
-			http.Error(w, fmt.Sprintf("error reading request body: %v", err), http.StatusBadRequest)
-			return
-		}
-		defer r.Body.Close()
-
-		fmt.Printf("Raw request body: %s\n", string(body))
-
-		var requestBody models.ModifyTaskRequestBody
-
-		err = json.Unmarshal(body, &requestBody)
-		if err != nil {
-			http.Error(w, fmt.Sprintf("error decoding request body: %v", err), http.StatusBadRequest)
-			return
-		}
-
-		email := requestBody.Email
-		encryptionSecret := requestBody.EncryptionSecret
-		uuid := requestBody.UUID
-		taskuuid := requestBody.TaskUUID
-		description := requestBody.Description
-		project := requestBody.Project
-		priority := requestBody.Priority
-		status := requestBody.Status
-		due := requestBody.Due
-
-		if taskuuid == "" {
-			http.Error(w, "taskuuid is required", http.StatusBadRequest)
-			return
-		}
-
-		if err := tw.ModifyTaskInTaskwarrior(uuid, description, project, priority, status, due, email, encryptionSecret, taskuuid); err != nil {
-			http.Error(w, err.Error(), http.StatusInternalServerError)
-			return
-		}
-
-		http.Redirect(w, r, "/tasks", http.StatusSeeOther)
-		return
-	}
-	http.Error(w, "Invalid request method", http.StatusMethodNotAllowed)
-}
-
-func AddTaskHandler(w http.ResponseWriter, r *http.Request) {
-	if r.Method == http.MethodPost {
-		body, err := io.ReadAll(r.Body)
-		if err != nil {
-			http.Error(w, fmt.Sprintf("error reading request body: %v", err), http.StatusBadRequest)
-			return
-		}
-		defer r.Body.Close()
-		fmt.Printf("Raw request body: %s\n", string(body))
-
-		var requestBody models.AddTaskRequestBody
-
-		err = json.Unmarshal(body, &requestBody)
-		if err != nil {
-			http.Error(w, fmt.Sprintf("error decoding request body: %v", err), http.StatusBadRequest)
-			return
-		}
-		email := requestBody.Email
-		encryptionSecret := requestBody.EncryptionSecret
-		uuid := requestBody.UUID
-		description := requestBody.Description
-		project := requestBody.Project
-		priority := requestBody.Priority
-		dueDate := requestBody.DueDate
-
-		if description == "" {
-			http.Error(w, "description is required", http.StatusBadRequest)
-			return
-		}
-
-		if err := tw.AddTaskToTaskwarrior(email, encryptionSecret, uuid, description, project, priority, dueDate); err != nil {
-			http.Error(w, err.Error(), http.StatusInternalServerError)
-			return
-		}
-		http.Redirect(w, r, "/tasks", http.StatusSeeOther)
-		return
-	}
-	http.Error(w, "Invalid request method", http.StatusMethodNotAllowed)
-}
